+++ bmc-filestorage/src/main/java/com/oracle/bmc/filestorage/model/ExportSet.java
@@ -249,7 +249,7 @@ public class ExportSet {
     Long maxFsStatBytes;
 
     /**
-     * Controls the maximum `ffiles`, `ffiles`, and `afiles`
+     * Controls the maximum `tfiles`, `ffiles`, and `afiles`
      * values reported by `NFS FSSTAT` calls through any associated
      * mount targets. This is an advanced feature. For most
      * applications, use the default value. The