+++ bmc-filestorage/src/main/java/com/oracle/bmc/filestorage/responses/UpdateExportResponse.java
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
+ */
+package com.oracle.bmc.filestorage.responses;
+
+import com.oracle.bmc.filestorage.model.*;
+
+@javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20171215")
+@lombok.Builder(builderClassName = "Builder")
+@lombok.Getter
+public class UpdateExportResponse {
+
+    /**
+     * For optimistic concurrency control. See `if-match`.
+     */
+    private String etag;
+
+    /**
+     * Unique Oracle-assigned identifier for the request. If
+     * you need to contact Oracle about a particular request,
+     * please provide the request ID.
+     *
+     */
+    private String opcRequestId;
+
+    /**
+     * The returned Export instance.
+     */
+    private Export export;
+
+    public static class Builder {
+        /**
+         * Copy method to populate the builder with values from the given instance.
+         * @return this builder instance
+         */
+        public Builder copy(UpdateExportResponse o) {
+            etag(o.getEtag());
+            opcRequestId(o.getOpcRequestId());
+            export(o.getExport());
+
+            return this;
+        }
+    }
+}